Thermal/requirements.txt
torch>=2.2.0
torchvision
gradio>=4.0.0
pandas
numpy
Pillow
diffusers
transformers
accelerate
lpips
tqdm
scipy
safetensors
matplotlib